restricting update and delete options only  to users who created posts

+++ backend/routes/posts.js
@@ -69,12 +69,18 @@ router.put('/:id',checkAuth,multer({storage:storage}).single("image"),(req,res,n
         _id:req.body.id,
         title:req.body.title,
         content:req.body.content,
-        imagePath:imagePath
+        imagePath:imagePath,
+        creator:req.userData.userId
     })
-    Post.updateOne({_id: req.params.id },post)
+    Post.updateOne({_id: req.params.id,creator:req.userData.userId },post)
     .then(result=>{
         console.log(result);
-        res.status(200).json({ message:'successfully updated',imagePath:imagePath});
+        if(result.nModified>0){
+            res.status(200).json({ message:'successfully updated'});
+        }
+        else{
+            res.status(401).json({ message: 'Not AUthorized'});
+        }
     })
 })
 
@@ -120,12 +126,15 @@ router.get('/:id',(req,res,next)=>{
 
 router.delete('/:id',checkAuth,(req,res,next)=>{
 
-    Post.deleteOne({_id:req.params.id})
+    Post.deleteOne({_id:req.params.id,creator:req.userData.userId})
     .then(result=>{
         console.log(result);
-        res.status(200).json({
-            message:"post deleted successfully"
-        }) 
+        if(result.n>0){
+            res.status(200).json({message:"post deleted successfully" }) 
+        }
+        else{
+            res.status(401).json({ message: 'Not AUthorized'});
+        }
     })
